Update inventory.md
My main purpose here was to try to reduce duplicates (in this case printers) and to update the camera to match its resolution to the device being tested instead of independent. I think the camera needs to be at least equivalent to the camera that would be used for the authentication.

I don't necessarily like the resolution listing on the laptop (that is a strange resolution, and not one I think I have seen).

What I am thinking is that we should standardize on needing a screen of specific resolutions and dimensions regardless of the form factor. It is possible that a laptop may meet some of the screen requirements (some tablets may do so as well), instead of listing them as laptop and PC screen.

I also wonder if we should split out the printers because I could see maybe wanting to have a test with an inkjet and a laser. I also think a photo printing service could be listed as a tool for consistency.