further updates
This is furst updates to the PPC and the SD for the alternative auth factor.

There is only one change in the SD.

The PPC has a mix of AAF and PAF because some of the references are to the MDFPP and others to the Bio, so these are kept separate and both are used in the appropriate place.

I also made some grammer updates to the PPC (mainly adding "the" into the document).