Update BiocPP.adoc
I observe that "Mobile Device" is used as example of TOE Environment. As Mobile device have their own definition as “such as smartphones, PC login at offices, ATMs at banks, and building or room entrance control, or border security checks.” in the ESR. Probably could be interesting to change it as “i.e. smartphone”. Apart of that, I believe that the first TOE reference is in deed TOE Environment. 
What did you think?

Finally there are something that I do not change but probably will be interesting to check. Is the "Mobile Device" Glossary definition. Actually appear as “A self-contained device which is composed of a hardware platform and its system software. The device is typically some sort of general purpose computing platform, such as a laptop, tablet or smartphone that is designed to be portable.” Anyway in the ESR “Use Case(s)” explain “Biometric products are used for user authentication for mobile devices such as smartphones, PC login at offices, ATMs at banks, and building or room entrance control, or border security checks.”
Could have sense to change it as “A self-contained device which is composed of a hardware platform and its system software. The device is typically some sort of general purpose computing platform, such as smartphones, PC login at offices, ATMs at banks, and building or room entrance control, or border security checks.”