removed Github links
I removed the GitHub links to the SD because it would cause problems in generating the final output and wouldn't be pointing necessarily to the "correct" location. When published, the docs would need to be pointing to the current version, not necessarily what is in GitHub.

I'm not sure exactly how to add another cross reference that would always work.