Editorial updates for proposed release
This is a set of editorial updates in preparation for the proposed release. All the docs have been set to a version of 0.95 ad a date of March 13.

Note that I changed "test document" in the BIOSD to "performance report" and linked it everywhere to that section except within that section (since that didn't seem to make sense)