Change password to Non-Biometric Authentication Factor
I have changed password and password authentication factor to non-biometirc authentication factor to handle the options that may be in use on PCs such as smart cards.

I didn't use straight up "authentication factor" since we had a specific definition for biometric authentication factor and technically AF would include biometric, so I specifically adjusted it to be knowledge/posession based.

This is to close #265