Create PP_Config-OS.adoc
Initial updated GPOS PP-Config.

This is based on the latest MDF version.

The main concern is section 5.4.1.5 which talks about the SEE. There isn't an SEE or equivalent required for the OS (which makes sense as it is not hardware). This could mean we can't make this work without more than some basic edits to the PP (i.e. auth edits to allow biometrics).

One possible method though would be to only allow approval based on some expectation that the biometric subsystem is isolated hardware and does not use the AP of the computer. So a self-contained biometric system using its own processor (I may say it could allow storage on the disk if it encrypts it first, like a TEE), would be the only acceptable devices that could be used here. A USB (or maybe BT) device would work, but not one where it handed everything off to the OS to process and store.

I'm not sure this will work, it may be layering too many requirements on how this should work.

I'm sending this version to Mike Grimm at Microsoft (since they are the company that validates end user OS, I don't know that Oracle would care about this topic).