Update for attack potential
This is to close #272 with the update to the attack potential table based on the updated categories from @n-kai

In addition I cleaned up the tables some, there there are a lot of changes, but the changes are mainly just moving the rows into separate lines so it is a little easier to read in the document. I also made a few minor edits to the formatting while cleaning up the tables.

The main changes are in lines 1236, 1318-1320, 1324 and 1622 (there are some other places in the tables where I added in the new content, but these are the main new pieces I have added).