Acronym update
This is to close #281 to add the acronyms to the table.

I also slightly edited the table to spread it out (i.e. instead of each row being on one line, I spread it out so each row is a set of lines with spaces and each column on a new row as is the case in most of the other tables).