Found "enroll"
The language for the PP-Module has used "enrol" but the definitions list has "Failure-to-Enroll". Given the standard usage of one "l", this should be edited.

This can wait until the next update and does not need to be made into an interpretation, but should be fixed in the next major release.