Consistency and Rationale edit
This is for @n-kai to show these commits directly.

This is based on the assumptions branch, but specifically moves his commits into a new branch (there is a little bit of a mess in the assumptions branch as I pulled them over).

This is to close #302 at this time.

There are changes here for #300 as well.