FPT_BDP_EXT updates
This closes #308 and closes #309.

This removes FPT_BDP_EXT.2/3 and replaces them with FPT_KST_EXT.1/2 from the MDF with modifications to add biometric data.

Secure Execution Environment -> Separate Execution Environment. This is from the MDF in sections FCS_CKM_EXT.1 & FCS_CKM_EXT.4, as a method of separation from the OS.

FPT_BDP_EXT.1 is changed to mirror FCS_CKM_EXT.1.2 in specifying some sort of hardware isolation (basically the SEE of some sort), and removed descriptions related to encryption for the EA.