FIA_MBV_EXT.1.1 remove assignment
This is to close #305.

This removes the assignment that we had allowed originally. The original thinking was that to use the assignment the vendor would have had to provide PAD and everything else, it wouldn't just be allowed, but NIAP prefers more explicit control (which is fine).

What I would expect as the proper course for this would be the vendor asks to add a new modality, and once everything has been submitted and approved, a TD would be issued to add the new modality to the list. This ensures that only approved modalities are allowed, and they must be approved before.

I alphabetized the list so there isn't any particular preference among the modalities.

I removed Voice because we don't have a toolbox for it at this time.

The other issue about the app note is already resolved by #315.