remove the app note
This is to close #307.

The other app notes mentioned are already removed due to other changes. It isn't clear that we actually need this app note though. While clearly it depends on the base PP, I don't know that it needs a note about consistency that doesn't actually provide information about the SFR itself.