Added BMD
Added biometrics management description based on the FE PP 1.0 from NIAP.

This is optional for the developer and so does not have the same expectations as the FE PP does in terms of the content.

This is to close #323.

The FE PP defines the expectations of the KMD in the SD, but since it isn't clear if every vendor will utilize the BMD, I do not think we can lay out requirements on the document or even in the SD as to what is expected where.