Close out TODO items
This is to try and close out the remaining TODO items for the MDF integrations.

The rationale table was added and the other places where something that was needed I have reviewed and edited as seemed necessary (in some cases there didn't seem to be anything that needed to be changed on review so the TODO was just removed).