editorial edits
While not necessarily everything, these are editorial edits I found while reviewing the BIOSD for the strict guidance.

All the ISO changes are around how it looks in the text so that everything looks like "ISO/IEC xxxxx" while still having the short codes as they were. This is more stylistic than anything else, but across the two docs to be consistent.