updates based on call
This updates to remove the memory test as out of scope as it is testing the SEE boundary, and not the TOE boundary.

In line 597 there is a statement (which I left) about source code review being acceptable to ensure this as an alternative to testing (which I left specifically for the TSFI test). I don't want to make this a requirement for the general boundary issue (that everything is inside the SEE), but it is something to consider.