Update BiocPP.adoc
This includes adding PAD-L to the glossary and updating the text to not say combinations are not public. As there is no mention of the toolbox in this doc, I didn't add anything there, but will update the website to show what is available.

includes a few changes of single quotes to ensure asciidoc output.