BMD conditional comment
This is to close #435. I did not see any other instances of this. Instead of making it conditional, since the information is required I wrote it that either the TSS or BMD has to include the information, so it is up to the developer to decide what to put where, but that it has to be available.