AGD guidance edit
This is to close #434.

I am actually deleting the section that caused the problem as I think it's more of a copy-paste error now that I review the content. I think the rest do still need to be modified in their placement as they are sort of hanging on at the end. This change moves all of them into the requirements as part of the bullet.

This specific bullet though is about selecting PAD, and doesn't actually have any AGD associated with it because the user isn't going to turn it on or off, so it shouldn't be there.